.github/scripts/requirements.txt
mail-parser
natsort
pyyaml
redbox
xmltodict
